chore(payroll): move date checks into datelogic.js
The manageConfig file contained lots of date logic that we should test.
I've decomposed it so that we can unit test some of the date logic and
spend time commenting the code instead of doing loops.

This is progress towards #3783.
